enviar email en laravel

+++ database/migrations/2020_04_18_094828_create_posts_table.php
@@ -20,13 +20,14 @@ class CreatePostsTable extends Migration
             $table->string('url_clean', 500)->nullable();
             $table->text('content')->nullable();
             $table->enum('posted', ['yes', 'not'])->default('not');
-            $table->integer('category_id')->unsigned()->nullable();
+            $table->bigInteger('category_id')->unsigned()->nullable();
             $table->timestamps();
-
+            /*
             $table->foreign('category_id')
                 ->references('id')
                 ->on('categories')
                 ->onDelete('cascade');
+                */
         });
     }
